pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dataspoc-pipe"
version = "0.1.0"
description = "Motor de ingestão de dados — Singer + Parquet + Bucket"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "DataSpoc", email = "dev@dataspoc.com" },
]
keywords = ["data", "ingestion", "singer", "parquet", "etl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]
dependencies = [
    "typer>=0.9,<1.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
    "fsspec>=2024.1",
    "rich>=13.0",
    "python-crontab>=3.0",
]

[project.optional-dependencies]
s3 = ["s3fs>=2024.1"]
gcs = ["gcsfs>=2024.1"]
azure = ["adlfs>=2024.1"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
dataspoc-pipe = "dataspoc_pipe.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/dataspoc_pipe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
